+++ RELEASENOTES.md
@@ -67,6 +67,8 @@
         ([#56](https://github.com/androidx/media/pull/56)).
     *   Fix RTSP basic authorization header.
         ([#9544](https://github.com/google/ExoPlayer/issues/9544)).
+    *   Throw checked exception when parsing RTSP timing
+        ([#10165](https://github.com/google/ExoPlayer/issues/10165)).
 *   Session:
     *   Fix NPE in MediaControllerImplLegacy
         ([#59](https://github.com/androidx/media/pull/59))